refactor(mcp-server): decompose app.ts god module into focused files
Extract utility helpers, type declarations, and server-mode middleware
from the 1051-line app.ts into three cohesive modules:

- app-helpers.ts (125 lines): errorMessage, isJsonObject,
  setBaselineSecurityHeaders, toInlineScriptJson, isReservedUiPath,
  SERVER_MODE_PLACEHOLDER_HTML, extractInitializeDebugPayload,
  decodeBranchTipOrThrow, checkoutCloneOrThrow
- app-types.ts (64 lines): LocalDaemonAppOptions, ServerModeAppOptions,
  AppOptions
- security/server-mode-middleware.ts (137 lines):
  createServerModeApiAuthMiddleware, createServerModeAsyncAuthMiddleware,
  createServerModeOriginMiddleware, sanitizeServerModeStatus

app.ts is reduced to 753 lines (the createApp function body and its
module-level wiring). Existing tests and type re-exports are preserved.
